MRD-2987 Move types to prison package
