+++ src/main/kotlin/uk/gov/justice/digital/hmpps/makerecalldecisionapi/domain/makerecalldecisions/SentenceSequence.kt
@@ -1,5 +1,7 @@
 package uk.gov.justice.digital.hmpps.makerecalldecisionapi.domain.makerecalldecisions
 
+import uk.gov.justice.digital.hmpps.makerecalldecisionapi.domain.makerecalldecisions.prison.Sentence
+
 data class SentenceSequence(
   val indexSentence: Sentence,
   val sentencesInSequence: MutableMap<Int, List<Sentence>>? = null,
